backend/html: outsource ToHtmlM type class
into seperate module
